+++ CHANGES.md
@@ -5,6 +5,7 @@ Version 0.8.5 (pending)
 - Improved signature pad detection
 - Fixed redirect to non-existent page when printing staff badges
 - Mark guests as checked in during attendee import
+- Add support for online purchases of single day badges
 
 Version 0.8.4 (10/26/2019)
 - Added attendance report