+++ CHANGES.md
@@ -1,5 +1,5 @@
 # Change History
-Version 0.8.5 (pending)
+Version 0.8.5 (11/8/2019)
 - MSO role permissions cleanup
 - Fixed pronouns in staff import
 - Improved signature pad detection